pyproject.toml
[project]
name = "genie-proxy"
dynamic = ["version"]
description = "Genie Proxy - Access Databricks Genie Spaces across workspaces with intelligent QPM request queuing backed by Lakebase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
]

[tool.apx.metadata]
app-name = "genie-proxy"
app-slug = "genie_proxy"
app-entrypoint = "genie_proxy.backend.app:app"
api-prefix = "/api"
metadata-path = "src/genie_proxy/_metadata.py"

[tool.apx.ui]
root = "src/genie_proxy/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.version]
path = "src/genie_proxy/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/genie_proxy/__dist__", "src/genie_proxy/_metadata.py"]
exclude = ["src/genie_proxy/ui"]
